Fixed cancel buttons, added new details view. Fixed styles for filter.

+++ php/controller/boatramp.php
@@ -25,14 +25,14 @@ class BoatRampController extends Controller
         /*** Instatiate a new boatramp model ***/
         $model = new boatramp();  
 
-
         /*** Get all Ramps ***/
         $ramps = $model->all();
-
         
         /*** set template variables ***/
         $this->registry->template->welcome = 'Boat Ramps';
         $this->registry->template->ramps = $ramps;
+        
+        
 
         
         /*** load the index template ***/
@@ -64,9 +64,41 @@ class BoatRampController extends Controller
 
         
         /*** load the index template ***/
-        $this->registry->template->show($this->name, 'new');
+        $this->registry->template->show($this->name, 'create');
     }
 
+    /**
+     * Loads the form to view a Boat Ramp
+     **/
+    public function view()
+    {
+    	/*** Instatiate a new boatramp model with the ID of the one we are editing ***/
+    	$this->model = new boatramp($_GET['id']);
+    
+    
+    	/*** Get the Boat Ramp where ID = model->id ***/
+    	$ramp = $this->model->at_id();
+    	/*** Get States ***/
+    	$states = array("NH","ME");
+    	/*** Get Towns ***/
+    	$towns = $this->getTowns();
+    	/*** Get Waterbodies ***/
+    	$waterbodies = $this->getWaterbodies();
+    
+    	/*** set template variables ***/
+    	$this->registry->template->welcome = 'View Boat Ramp';
+    	$this->registry->template->states = $states;
+    	$this->registry->template->towns = $towns;
+    	$this->registry->template->waterbodies = $waterbodies;
+    	$this->registry->template->ramp = $ramp[0];
+    
+    
+    	/*** load the edit template ***/
+    	$this->registry->template->show($this->name, 'view');
+    
+    }
+    
+    
     /**
      * Loads the form to edit a Boat Ramp
      **/
@@ -115,24 +147,66 @@ class BoatRampController extends Controller
     * Gets called when an Edit Boat Ramp form is submitted
     **/
     public function update() {
-    	if (isset($_SESSION['roleID']) && ($_SESSION['roleID'] < 3)) {
-	        $model = new boatramp($_GET['id']);
-	        $model->updateBoatRamp($_POST["ramp"]);
+    	$edit = $_POST['edit'];
+    	 
+    	if($edit == "Submit")
+	    {
+	    	if (isset($_SESSION['roleID']) && ($_SESSION['roleID'] < 3)) {
+		        $model = new boatramp($_GET['id']);
+		        $model->updateBoatRamp($_POST["ramp"]);
+	    	}
     	}
-    	
     	/*** Redirect User to BoatRamp/Index ***/
     	header("location: index.php?rt=boatramp/index");
     }
     
+    
+    /**
+     * Loads the form to view a Boat Ramp to delete
+     **/
+    public function delete()
+    {
+    	/*** Instatiate a new boatramp model with the ID of the one we are editing ***/
+    	$this->model = new boatramp($_GET['id']);
+    
+    
+    	/*** Get the Boat Ramp where ID = model->id ***/
+    	$ramp = $this->model->at_id();
+    	/*** Get States ***/
+    	$states = array("NH","ME");
+    	/*** Get Towns ***/
+    	$towns = $this->getTowns();
+    	/*** Get Waterbodies ***/
+    	$waterbodies = $this->getWaterbodies();
+    
+    	/*** set template variables ***/
+    	$this->registry->template->welcome = 'Delete Boat Ramp';
+    	$this->registry->template->states = $states;
+    	$this->registry->template->towns = $towns;
+    	$this->registry->template->waterbodies = $waterbodies;
+    	$this->registry->template->ramp = $ramp[0];
+    
+    
+    	/*** load the edit template ***/
+    	$this->registry->template->show($this->name, 'delete');
+    
+    }
+    
+    
     /**
     * Gets called when a Boat Ramp is deleted
     **/
-    public function delete() {
-    	if (isset($_SESSION['roleID']) && ($_SESSION['roleID'] < 3)) {
-	        $model = new boatramp($_GET['id']);
-	        $model->deleteBoatRamp();
-    	}
+    public function deleteRamp() {
     	
+    	$delete = $_POST['delete'];
+    	
+    	if($delete == "Delete")
+    	{
+	    	if (isset($_SESSION['roleID']) && ($_SESSION['roleID'] < 3)) {
+		        $model = new boatramp($_GET['id']);
+		        $model->deleteBoatRamp();
+	    	}
+    	}
     	/*** Redirect User to BoatRamp/Index ***/
     	header("location: index.php?rt=boatramp/index");
     }
